fix: moved getMaximumNumbersLength in kup-math
